fix(chisel): allow same ip address for p2p conns
fixes #475.

a: self validator
b: other validators

we run a Chisel client for each (b) in the same host where we run (a).
while doing so, every (b) gets same(127.0.0.1) ip address -but with a
unique port number.
this PR enables allowing to use same ip address for each (b).