feat: add pkg/protobufjs
protobufjs is made available for use as a Go pkg with the high level Generate() API.

protobufjs binaries statically added to Starport's source code and they are instantly available.

only the relavent protobufjs binary included to the final binary of Starport dependending on the OS. for Linux, it adds 17MB to the total binary size of Starport.