feat: don't require GOPATH/bin or GOBIN env vars to be set
add a small trick to automatically detect installation dir for go
programs and set it to the env while running commands.

this is specially relavent with protoc plugins because they are
installed with `go get` on the fly and persisted  under the installation
dir.

not configring the installation dir (adding $GOPATH/bin or $GOPATH to the $PATH)
means incomplete installation of Go but this will not be a blocker for Starport anymore.