feat(xrelayer): wire Go xrelayer with TS relayer wrapper
`starport/pkg/xrelayer` previously was using go-relayer to implement a high level relayer wrapper/abstraction.
as we switch to the ts-relayer, this abstraction layer will be implemented in TS under `scripts/data/gen-nodetime/src/relayer` for `starport/pkg/xrelayer` to use in Go.
this commit establishes the wiring between the Go xrelayer and TS wrapper.

see `scripts/data/gen-nodetime/src/relayer/lib` to implement functions that are required by the Go xrelayer in TS.

Starport developers needs to execute `scripts/gen-nodetime` when there is change in the impl of the TS wrapper so, Go code can access to the updated nodetime binary with these changes.

wiring is done with a basic subset of json-rpc 2.0.

* done some refactoring on nodetime.