feat(build): 🔨 Introduce Taskfile as a replacement (or alternative) for Makefile
Hello! This PR (and a few others I'll be submitting) is part of a brief
conversation I had in the hackatom channel on discord. I made these
changes for myself for the HackAtom VI hackathon, and figured others
would get an ease of use out of the tool.

While this [Taskfile](https://taskfile.dev) results in a larger set of
build files for building starport, it also allows to keep things more
compartmentalized and error proof. Few changes have been made at the
moment beyond the following:

1. Targets like `test-integration` and `test-unit` are now
   `test:integration` and `test:unit`.
2. Some shell commands have been removed, reducing shell output.
3. Not all targets are silent by default.
4. The `Makefile` has not been deleted.

This change also brings the following features:

1. Users can run `task --list` to get a list of all targets
2. A `bootstrap` target has been added to ensure that all *tools*
   required for editors + plugins like vscode-go or vim-go can be
   installed. This target also installs `delve`, the powerful golang debugger.
   This target is optional to call, but also means that users simply
   have to have the Taskfile tool installed before actually getting
   started with development.
3. Running `task --summary <task>` will sometimes give a detailed
   summary of what the target does or its configuration settings.
4. To get started with *developing starport itself* users only need to
5. Though not yet implemented in the `format` task, the more powerful
   and strict `gofumpt` tool is available to install via
   `bootstrap:gofumpt`.

Lastly, I do not expect this to be merged immediately as there are one
or two style pieces as well as some possibly inaccurate docs. I've also
not updated the github workflows, readme, or other documentation but am
committed (pun not intended) to pushing these changes before merge.