Run go install after go get to install plugins
This fixes https://github.com/tendermint/starport/issues/1163, but is not ideal... tooling binaries should be installed by running `go install [path]@[version]`, but protoc-gen-gocosmos uses a replace directive in the `go.mod` which is [incompatible with this method](https://github.com/golang/go/issues/44840#issuecomment-792789581). This fixes the problem for now, but adds plugin specific items to the go.mod of the scaffolded project, which shouldn't be there.