refactor: replace owner name references with app module paths
The owner name shouldn't be used anymore because the Go module paths
might be defined as a single name or a URI with a single path element
that only describes the app name, so in those cases there is no owner
definition.

App module paths are used instead of the owner name.

App module paths are extracted from the Go module path, e.g.:
 - github.com/username/reponame => "username/reponame"
 - domain.com/reponame => "reponame"
 - reponame => "reponame"
 - namespace/reponame => "namespace/reponame"
 - a/b/c/namespace/reponame => "namespace/reponame"