refactor: change TX to read parsed events
The refactor was required because the transaction log might contain
plaintext in some cases, for example when there are not enough funds to
pay the transaction fees.

The parsed Tendermint search result events are used instead. These
events contain JSON encoded values that are unquoted, including the
string values, so all event values should be parsed to fix this
inconsistency.