feat(pkg/cosmosmetric): add support for managing separate SQL schemas
This feature lets the postgres adapter keep the schema versioning
functinality while allowing external apps to also keep track of their
schemas. This is useful to let the apps create views or functions to
improve how TX and events are queried moving the loginc to the database
instead of the app.