fix: change `protoc-gen-ts_proto` to generate in random folders
The script has a fixed name to comply with `protoc` plugin naming
requirements but it it always generated in the same temporary folder
which means that concurrent calls would overwrite the `protoc-gen-ts_proto`
file.

This changeset generates the script in random temporary folders.

This fix might help correcting "flaky" integration tests.