Ensure bank send can take account name or address
Also simplify the logic by removing the read of the --from flag, which
is no longer used in this command.

BroadcastTx now takes a cosmosaccount.Account instead of just an account
name. Since ensure the account comes from the keyring, and avoid the
BroadcastTx impl to re-check that (because the keyring is always
checked before the call to BroadcastTx).