chore: add go formatting github action (#2774)
* chore: add go formatting github action

- Use goimports with `-local` flag which separates ignite/cli package from
the others.
- Use gofumpt instead gofmt

The gh action is triggered when a commit with go file change is pushed to
develop. If goimports or gofumpt have changed some files, then an other commit
is pushed to develop. Nothing happens if there's no changes.

I initially triggered the action on pull requests, but I found annoying
and intrusive to have these new commits in the context of a feature
branch.

Note this commit contains an unrelated change to confile_test.go, in
order to test that gofumpt will fix that. For goimports there's already
a couple of changes that will happen.

* centralize formatting in makefile

also add versionning to formatting tools

* use go run package directly

* create/update a PR rather than committing to develop