fix(cosmosclient): call the faucet prior to create the tx
Fix #2775

This change restores the behavior we used to have in 0.23, and which has
changed with the refac of cosmosclient with the `node` command feat.

Calling the faucet (when it is specifically enabled) will ensure the
account that broadcasts the tx exists in the target chain.

Because the `node` command has a `--generate-only` flag, and because we
don't want to create a new account when it's set, this flag is now
propagated to the cosmosclient struct. Therefore, the faucet isn't
called when the flag is set. This doesn't affect the `network` command
because it doesn't use that flag.