refactor: use struct to manage config's server addresses
Map type causes a lot of trouble and extra casting and checks when the
Cosmos SDK or Tendermint config values for the server addresses are read
or modified so instead of that this changeset uses a struct.
We can't use Cosmos SDK ot Tendermint config structs because if any of
there configs gets updated it could break the migrations.

The changeset also fixes the default address values that were encoded to
the YAML config. These values no longer appear if they are the default
ones.