fix: change init chain to support merging YAML generated maps
The contributed implementation was removing some fields which was not
correct; Now those config fields are saved during init.
It seems that the map fields were being removed to avoid an issue with
maps containing interface{} key types instead of string.
This change fixes it by saving all the config values and by adding
support for YAML generated maps thought a transformer implementation.