refactor: change chain's init to update app config files only once
The blockchain app config file were updated twice, once by the stargate
plugin and then again by the chain init.
This changeset updated the app's config only once. The stargate plugin
is the one that updates them.
The changeset also solves the issue where come of the addresses that
must have the protocol prefix were ovewritten with an address without
prefix because the host config value was removed (#2470), so now
addresses are defined in the app and config sections.