chore: set union merge strategy to changelog.md
Since we are forcing the updage of changelog.md on every PR, this leads
to many conflicts when it's time to merge.

By forcing the merge strategy of this file to `union`, it decreases the
chance of conflicts by ~90%. The downside is sometimes the merge gets
wrong (repeated lines) so the dev should have a look at the result.