test: TxService.Broadcast
The change introduces a new interface to mock the call to tx.Sign(). I
found more interesting to assert the call to that function rather than
trying to assert its behavior in term of code, which is basically
testing the cosmos-sdk, and we don't want to do that.

By adding the test I found a bug in the error comparison that I
introduced previously with the rpcWrapper, so I fixed it.