fix: change nightly tag format
Fix #2811

This changes the nightly tag format to `nightly`. This is the only
change, and that should mitigate the errors reported in the issue.

Previous format v0.0.0-nightly was registered in sum.golang.org with the
old org `ignite-hq/cli`, which was triggering a checksum error when doing
`go get` with the new org.

Additionally, using a non-semantic tag is probably preferrable to
prevent the situation from happening again in the future.

It's still possible to download the nightly version of ignite using the
official website, by requesting `https://get.ignite.com/cli@nightly`
(previously it was `https://get.ignite.com/cli@v0.0.0-nightly`).