chore: create openapiconsole sub module
Part of #2658

This is the first step of releasing openapiconsole as a separate go
module inside the cli repository, which consists of:
- adding a go.mod in ignite/pkg/openapiconsole/ folder
- declaring openapiconsole dep in the main go.mod (because it's currently
  used by pkg/cosmosfaucet)

The second step, once this PR is merged, is to push a specific tag for
this new module, which must look like `ignite/pkg/openapiconsole/v1` (I
think using `v1` is OK for such a small module).

One the tag exists, a second PR will remove the `replace` directive and
change the dep version for this tag.