feat: handle module discovery for akash
Fix #2864

To discover akash's modules, the module discover needed to handle:

1. call to `module.NewBasicManager()` in a different file of `app/app.go`:
This is done by parsing the whole app package instead of only the app.go
file.

2. some modules are returned by a call to local function:
This is done by handling more scenarios when iterating over
`module.NewBasicManager()` arguments

3. `module.NewBasicManager()` arguments are passed via the `append`
   function:
This is done by manually checking if one of the argument is the `append`
function.

The change also comes with a refactoring:
- improve error handling
- introduce xast package