feat: plugin system
A plugin is a binary that can communicate with ignite using
github.com/hashicorp/go-plugin. When ignite is executed under a chain
configuration, it checks if the config contains plugins. If yes the
plugins are fetched, compiled and ran, which will add more commands to
the existing ones.

Fetching and compilation are cached under .ignite/plugins folder.

A new command `ignite plugin` has been added to manage plugins and make new
ones.

To give plugins the ability to load the chain, the function
`cmd.NewChainWithHomeFlags` is now public.