chore: remove get-diff action
This action has probably been introduced to skip longer action like lint
and test when there's no change. Unfortunately now the codebase is too
large (probably because of binary files), and the action is now longer
than lint and tests.

Instead of trying to tune it to skip binary files, this change simply
removes it, because lint and test actions aren't that long anyway.