feat: plugin system (rebased) (#2877)
* feat: plugin system

A plugin is a binary that can communicate with ignite using
github.com/hashicorp/go-plugin. When ignite is executed under a chain
configuration, it checks if the config contains plugins. If yes the
plugins are fetched, compiled and ran, which will add more commands to
the existing ones.

Fetching and compilation are cached under .ignite/plugins folder.

A new command `ignite plugin` has been added to manage plugins and make new
ones.

To give plugins the ability to load the chain, the function
`cmd.NewChainWithHomeFlags` is now public.

* update plugin go.mod

use hashicorp/go-plugin 1.4.4 to skip warning message

* add plugin go.sum

* git init and commit scaffolded plugins

* update CL

* fix dependecy issue

* fix missing commit author when no default config

CI doesn't have any default git config.

* fix linter

* Update ignite/services/plugin/template/main.go.plush

Co-authored-by: Alex Johnson <alex@shmeeload.xyz>

* remove plugin go.sum and use gocmd

* Update ignite/services/plugin/plugin.go

Co-authored-by: Jerónimo Albi <jeronimo.albi@tendermint.com>

* remove commented line

* fix cannot use cmd.Context() there bc it's nil

* kill plugin client at the end of execution

* fix comment

* improve error handling in LoadPlugins

* remove plugin.Name

The field has no real usage, it was here to have a shorter name than the
path but dealing with the path is actually less confusing.

* fix readme for plugins

* fix: ensure plugin instances are killed

* update main so defer funcs are honored

* use const exit codes

* go mod tidy

* fix due to plushgen removal

* add PR id to CL

* change CL version

* fix CL

Co-authored-by: Alex Johnson <alex@shmeeload.xyz>
Co-authored-by: Jerónimo Albi <jeronimo.albi@tendermint.com>