fix test script
When no match, grep returns 1, when xargs command return an error, xargs
return 123. Because of `set -e`, all commands must return 0, which is
good IMO. The alternative is to just check $? after the `go test`, but
then some errors would be ignored in the remaining of the script.

Also fix the test fix on `config/plugins.ParseDir`. There's no need to
have an empty file in the empty dir.