fix network integration tests
The network integration tests used to fetch the SPN version from CLI
go.mod, but this is no longer possible since the dependency has been
removed.

To fix that, the SPN version is now read from the cli-plugin-network's
go.mod which is cloned from the declared version in defaultPlugins.

A fix has also been added to support persistant flags.

Finally pluginsconfig.HasPlugin has been updated so it can return true
if a local plugin with a module name equals to the given path exists.
This permits to use `ignite` with a local cli-plugin-network.