feat: new ignite doctor command
Fix #3291

- add ignite doctor command (hidden for the moment)
  - migrate config.yml if needed
  - create tools/tools.go if not present, then install dependency tools.
    The `tools/tools.go` scaffold is handled by the same generator as
    `scaffold chain`, with an update allowing it to filter the fileset.
    `xgenny.NewEmbedWalker`, which was used to remove the `files/`
    prefix, is replaced by `fs.Sub` from the standard library.
- update `cosmosgen.InstallDepTools()`
  - remove retro compatibilty with chain that doesn't have
    `tools/tools.go`, in other words, remove the call to `go get`
  - run `go mod tidy` to update `go.sum`
  - suggest to run ignite doctor when dependency tools install fails
- move `GOSUMDB=off` inside `gocmd.ModTidy` instead of setting it prior
  to the call. This ensures it's always set.

Because the command uses intensively the filesystem, I chose to test it
with integration tests only. The integration test introduces a new library
called `testscript` [0], which is very convenient for this kind of tests.

`testscript` is based txt files, where each file is a test case that
contains some instructions. The txt file for `ignite doctor` runs
`ignite doctor` and then asserts what has changed in the filesystem,
according to some commands like `cmp`. The initial state before the
command is run is also specified in the txt file, with a list of files
designated by `-- filename --` followed by the file content.

[0] https://pkg.go.dev/github.com/rogpeppe/go-internal/testscript