docs: add migration steps for Ignite CLI v0.27.1 (#3558)
* docs: add initial migration steps for v0.27.1

There are more `app.go` changes to be added to the migration guide and
after that the document must be copied to nightly docs.

* docs: add `app.go` changes

* chore: remove init function from app changes

* chore: add missing remove line

* docs: add missing migration steps

* docs: add app changes for IBC migration

* docs: add missing migration steps

* docs: add missing migration steps

* chore: add link to parameters migration

Using the new Ignite CLI version won't work with an existing app state
unless the parameters are migrated. Without the migration the app can
oly be run using the `--reset-once` flag.

* chore: copy migration docs to nightly documentation

* docs: correct issues in documentation

* fix: correct link issues

The issues have no relation to the migration PR but fixes are required
to pass CI checks.