feat: re-enable TS codegen (#3655)
* feat: Re-introduce module discovery

* feat: Look for indirect proto deps in go.mod

* feat: Introduce indirect proto dependency discovery

* feat: Add buf export command

* feat: User buf-based openapi generation

* fix: Revert to previous sta version & peg TS version

* feat(wip): Update dependency resolution algorithm

* feat: Clean up include resolution

* chore: Add changelog entry

* chore: Fix linting issues & typos

* fix: cosmos-sdk buf issue

* feat(pkg/protoc): change package to use protoc binary from files repo (#3657)

* feat(pkg/protoc): change package to use protoc binary from files repo

* chore(pkg/protoc): remove embedded protoc binary

* fix: correct buf.gen.sta.yml file

* fix: address review

* fix: Remove logging

* fix: ResolveDependencies call

* fix: correct lint issues and merge `main` (#3690)

* chore: remove redundant variable

* chore(pkg/cosmosbuf): improve snippet redability

* chore: remove redundant default value

* chore: code formatting

* feat(pkg): remove nodetime binaries (#3670)

* - remote nodetime binaries
- import ignite-files repo pkg
- remote gen binaries scripts and pipeline
- run gofmt

* remove tollchain from go.mod

* add changelog

* Update changelog

* ci: update Go version to `1.21` in GitHub workflows (#3687)

---------

Co-authored-by: Pantani <Pantani>
Co-authored-by: Jerónimo Albi <jeronimo.albi@tendermint.com>

* fix: change Go version in nightly release config to `1.21.2` (#3688)

* fix: change Go version in nightly release config to `1.21.2`

The release action requires a version that includes the patch version to
be able to download the right tarball with the Go binary.

See https://github.com/wangyoucao577/go-release-action/blob/2ac3035fa4c4feed6a8272ce278b0577b93cf8e5/setup-go.sh#L24

* chore: change go releaser to use the lates Go version

Latest is the default but it's setted here to be explicit about that fact.
See https://github.com/wangyoucao577/go-release-action#parameters

* fix: remove explicit "latest" for Go version release nightly (#3689)

Lates version is used by default but it shouldn't be specified
explicitly because the value "latest" is not supported.

* chore: merge `main`

* test: fix broken integration test

* fix: change scaffolder to skip protoc when no module is scaffolded

There are no proto file when an app is scaffolded without module which
makes buf export fail because buf workspace references a folder without
proto files.

* fix(pkg/cosmosbuf): add check for proto files before running Buf

Changes previous commit because some files like the OpenAPI file should
be generated even when there are no proto files. We should also allow
generating code for standard dependencies like Comos SDK when the app
doesn't have proto files.

---------

Co-authored-by: Danilo Pantani <danpantani@gmail.com>

* feat: Bump nodetime version

* refactor: cleanup tmp folders & change third party includes (#3696)

* fix: implement third party includes using previous semantics

Use `thirdModuleIncludes` instead of defining a new type to store
modules and includes to be consistent with `appIncludes` and to avoid
changing existing code to work with a new type.

* feat: add tmp dir cleanup support to cosmos generator

* chore: restore Go version to `1.21` and remove toolchain

Currently CI is giving an error because is not properly configured to
support the new Go 1.21 features for the mod file and we don't have a
consensus yet on how to configure the toolchain and Go versions.

* chore: go mod tidy

* ci: change integration tests to use Go version from mod file

This is required to avoid downloading a newer Go toolchain which will
fail because GOSUMDB is disabled to fix the timeout issues because of
the repository size.

* ci: disable GOTOOLCHAIN when running integration tests

This is required because toolchain won't work when GOSUMDB is disabled.

* ci: change integration tests to use the stable go version

The stable Go version should be used to avoid keeping updating the
workflow config when a new patch version is released. This is to make
sure that Go doesn't try to download a new toolchain version.

* fix: correct CI linting issues

* ci: remove redundant GH workflow cache step

Go setup action caches since `v4`

* fix: disable go setup GH action cache

Action for `golangci-lint` already have caching functionality so Go setup
must be disabled to avoid caching issues.

* chore: add thitd party includes only when available

* fix: correct directory remove

Co-authored-by: Danilo Pantani <danpantani@gmail.com>

* fix: Address code review comments

* chore: Address review comment

* chore: remove commented function

* Update ignite/pkg/cosmosbuf/buf.go

Co-authored-by: Jerónimo Albi <jeronimo.albi@tendermint.com>

* chore: fix function name

* Update ignite/pkg/cosmosgen/generate_openapi.go

Co-authored-by: Danilo Pantani <danpantani@gmail.com>

* chore: address review

* chore: linting

* chore: test --experimental_allow_proto3_optional flag

---------

Co-authored-by: Jerónimo Albi <jeronimo.albi@tendermint.com>
Co-authored-by: Danilo Pantani <danpantani@gmail.com>
Co-authored-by: jeronimoalbi <jeronimo.albi@gmail.com>