refactor(templates): introduce alternative folders (#3897)
* refac(templates): introduce alternative folders

Continuing @julienrbrt's work with the minimal chain feat, this change
updates a little bit the way some template can be overrided.

Initially, there was an alternate file `app_config_minimal.go` that
existed belong to the other files, and this file had to be excluded by
default.

I suggest to introduce an alternative using folders, aside the initial
`files`. So for the minimal chain, there's now a new `files-minimal`
folder that contains all the files that will be overrided from the ones
in `files`, when a minimal chain is scaffolded.

This simplify the include/exclude logic, which could become very
cumbersome if we need to introduce a second alternative like for the
consumer chain. This also avoid conflicts.

The rule to follow in my opinion is, when the templates contains too
many `<%= if IsCondition >`, then you should probably introduce an
alternative `files-*` folder. This is exactly what I'm going to do for
the consumer chain (see #3660).

* add missing CL

* clean unused imports