feat(plugin): execute from func
This change introduces a `plugin.Execute()` function that can be used to
run arbitrary code without affecting the Ignite CLI dependencies. The
code lives in its own repo, like any other plugins. Arguments can be
passed to update the behavior.

Once merged, this will be used to remove the interchain-security dependency
from Ignite CLI which we currently have in #3660 (see discussions [0])

Because this will be used internally in Ignite CLI for specific cases,
I don't feel that we need to document this feature in the official
documentation.

[0]: https://github.com/ignite/cli/pull/3660#discussion_r1448481624