refactor(services): add suggestions for logging on `serve` failure (#3943)
Currently when `serve` fails, it shows
```
Blockchain failed to start.                           
If the new code is no longer compatible with the saved
state, you can reset the database by launching:       
ignite chain serve --reset-once         
```

Considering this will the entry point for many folks new to Ignite, probably adding a suggestion for verbose logging is a trivial, but helpful addition.

Cheers.